+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Bug Fixes
 
+* **components:** 百度地图 ([d2d36cd](https://github.com/dgiot/dgiot-dashboard/commit/d2d36cd565d8dde402892789ce01a84aa47f0752))
 * **components:** 设备列表 ([7a5b59b](https://github.com/dgiot/dgiot-dashboard/commit/7a5b59b9bbbd9163e5df359224f5b6b433b2432e))
 * **components:** 设备列表 ([a9dd45c](https://github.com/dgiot/dgiot-dashboard/commit/a9dd45ca5044fa0532807a114feab6d8125d625b))
 * **components:** 设备列表 ([36e2951](https://github.com/dgiot/dgiot-dashboard/commit/36e29513ec2e8a88ccb586e1779800f81aa4b771))